# specs/swaggerdoc.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Medten Product API'
  description: 'Interact with the Medten product database using this API.'
servers:
  - url: https://mgtwitxxi1.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API Test Environment
  - url: https://v73kddyeqk.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API POC (Proof of Concept) Environment
  - url: http://localhost:3000/api/v1/
    description: Developer Local Test Environments
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/gaumeister/medten-product-api/1.0.0
paths: 
...
  /dataload/tables:
    get:
      summary: get the list of table entity names for use in the other dataload API calls.
      tags:
        - dataload
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items: 
                      type: string
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
  /dataload/${entity}/metadata:
    get:
      summary: get table metadata (field and data-type information).
      tags:
        - dataload
      parameters:
        - in: path
          required: true
          name: entity
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: entity name of the table
                  plural:
                    type: string
                    description: plural entity name of the table
                  table:
                    type: string
                    description: actual db table name
                  columns:
                    type: array
                    items:
                      type: object
                      properties:
                        column:
                          type: string
                          description: name of the db column
                        sql_type:
                          type: string
                          description: the data type (including length, precision) of the column
                        pk:
                          type: boolean
                          description: whether the column is/is part of the primary key
                        nullable:
                          type: boolean
                        default:
                          type: string
                          description: the default value, if any
  /dataload/${entity}/bulkupdate:
    post:
      summary: mass update data ina a table
      tags:
        - dataload
      parameters:
        - in: path
          required: true
          name: entity
          schema:
            type: string
      requestBody:
        description: The data to be loaded. The expected content is a tab-delimited list of data to load. The data should contain a first row matching table columns (see the `/dataload/${entity}/metadata` API for these details). Column(s) matching the primary key of the table **must** be provided to establish a match. All data provided will be updated. Data that does not match a row in the database will be ignored.
        required: true
        content:
          text/plain:
            schema:
              type: string
            example: id\tname_en\tname_zh\tparent_id\tis_oem\n1\tGE Healthcare\t通用\t0\t0\n2\tMindray\t迈瑞\t0\t0
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  total:
                    type: integer
                  inserted:
                    type: integer
                  skipped:
                    type: integer
                  updated:
                    type: integer
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /dataload/${entity}/bulkinsert:
    post:
      summary: mass insert data into a table
      tags:
        - dataload
      parameters:
        - in: path
          required: true
          name: entity
          schema:
            type: string
      requestBody:
        description: The data to be loaded. The expected content is a tab-delimited list of data to load. The data should contain a first row matching table columns (see the `/dataload/${entity}/metadata` API for these details). All data provided will be inserted, regardless of whether matching data exists.
        required: true
        content:
          text/plain:
            schema:
              type: string
            example: id\tname_en\tname_zh\tparent_id\tis_oem\n1\tGE Healthcare\t通用\t0\t0\n2\tMindray\t迈瑞\t0\t0
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  total:
                    type: integer
                  inserted:
                    type: integer
                  skipped:
                    type: integer
                  updated:
                    type: integer
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /dataload/${entity}/download:
    get:
      summary: download the contents of a database table (up to 100,000 rows)
      tags:
        - dataload
      parameters:
        - in: path
          required: true
          name: entity
          schema:
            type: string
      requestBody:
        required: false
        content: {}
          
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching entities in the database.
                    type: integer
                  plural_entity_name:
                    description: the `plural_entity_name` property will be the pluralized name of what is being downloaded ('products', 'families', 'equipment', etc.).
                    type: array
                    items:
                      type: object
                      description: the individual objects data corresponding to rows in the table.
                  headers:
                    type: array
                    items:
                      type: object
                      description: a single object whose properties can be used to create headers for CSV/table rendering.
            text/csv: {}
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
        error:
          type: string
          description: The error details. Typically error details may contain sensitive information not intended to be shown directly in the user interface.
...
  responses:
    400ClientError:
      description: client error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'
...
    500ServerError:
      description: server error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'